requirements.txt
absl-py
altair
astunparse
attrs
blinker
cachetools
certifi
charset-normalizer
click
Flask
flatbuffers
gast
gitdb
GitPython
google-pasta
grpcio
h5py
idna
itsdangerous
Jinja2
jsonschema
jsonschema-specifications
keras
libclang
Markdown
markdown-it-py
MarkupSafe
mdurl
ml-dtypes
namex
narwhals
numpy
opt-einsum
optree
packaging
pandas
pillow
protobuf
pyarrow
pydeck
Pygments
python-dateutil
python-dotenv
pytz
referencing
requests
rich
rpds-py
six
smmap
streamlit
tenacity
tensorboard
tensorboard-data-server
tensorflow
tensorflow-io-gcs-filesystem
termcolor
toml
tornado
typing_extensions
tzdata
urllib3
Werkzeug
wrapt
